Add files via upload
These are zylab_10.19 and 11.22 from Homework 3. I was able to pass the first couple of tests but when the menu is implemented the second time the output is not expecting user input to be printed. If I remove this piece the first several tests fail but I can get points for the  latest section. I am at a loss on how to correct this and I want feedback as to how to fix it. 

Thanks,